Added a Separate Measurement Register containning Exclusively the Measurements Outcomes
Renamed the old Binary Register as Prediction Register.
The Measurement Register is updated only when Measurement is Performed.
The Prediction Register has the Same Behaviour.
Adapted the "display" Instruction to the new Changes: both Prediction and Measurement Registers are Displayed Separately.
Fixed a Bug in the Depolarizing Noise implementation: Missing Gates in the generated Noisy Circuit.