generic_worker.h: remove duplicate unlock
Unlocking an unlocked mutex results in undefined behaviour according
to the man page. Remove the logic error that would result in calling
unlock twice in a row on the same mutex. This solves issue #35.

I'm not too thrilled about the try/catch block in the run() method
either, but i'll create a ticket for that. In general, the multi-
processing code should be checked for logic errors like this.