Update lockable_factory.cc
The class `mutex` from the nested namespace `xpu::core::os` can lead to lookup ambiguity with the class `mutex` from the namespace `std::__1::mutex` if QX is used as API within a C++ code that indirectly includes the `mutex` class from the standard library. As example, this happens of QX and OpenQL are included as API in the same C++ code.