Fix issue #114
The multithreaded/intrinsics version of the measurement operation seems
to contain bug(s).

The sequential version works, as proven by the various unit tests added
in this change.

This change disables the multithreaded measurement gate in favor of the
working single-threaded version. Bugs in the parallel implementation can
be investigated separately, when time allows.