Change Matrix and DenseUnitaryMatrix to use vectors instead of arrays. This way, they don't have to be templated anymore. Add default_gates map. With these changes we can get rid of the if-else block in get_gates. Most importantly, we can implement gate modifiers by just changing the unitary matrix. Change Unitary to hold pointers to both matrix and operands.
Add a set of functions to operands_helper.cpp to unroll a list of operands (SGMQ groups) to a list of instructions indices.

Change x > EPSILON comparisons to x >= EPSILON.
Change BitIndex and QubitIndex to aliases of Index.

TODO: check if we can build a UnitaryMatrix without specifying Matrix and Row.
TODO: add gate modifier tests.